properly deal with (practically) unlimited cache

+++ crates/polars-lazy/src/physical_plan/executors/cache.rs
@@ -5,7 +5,7 @@ use super::*;
 pub struct CacheExec {
     pub input: Box<dyn Executor>,
     pub id: usize,
-    pub count: usize,
+    pub count: u32,
 }
 
 impl Executor for CacheExec {
